requirements.txt
--index-url https://download.pytorch.org/whl/cu126
torch==2.9.0+cu126
torchvision==0.24.0+cu126
torchaudio==2.9.0+cu126
Pillow==11.3.0
ipython==9.6.0
ipywidgets==8.1.8
matplotlib==3.10.7
nibabel==5.3.2
pandas==2.3.3
psutil==7.1.3
scikit-learn==1.7.2
scipy==1.16.3
timm==1.0.21
tqdm==4.67.1
pyyaml==6.0.3
